refactor: change and update various classes
- Updated `SuDungDichVuRequest`, `SuDungDichVuResponse`, and `SuDungDichVu` classes to change the data type of consumption fields from BigDecimal to Integer.
- Enhanced `HopDongPhongRequest`, `HopDongPhongResponse`, and `HopDongPhong` classes to include new service fields (dvRac, dvWifi, dvCap, dvKhac).
- Modified `HopDongPhongService` to accommodate new service fields in request handling and response formatting.
- Removed obsolete `KhachThue` and `TrangThai` entity classes and their associated repository.
- Updated binary class files for various controllers, services, and DTOs to reflect the above changes.

+++ SoTroOnline/src/main/java/com/so_tro_online/app/QuanLyHoaDon/src/main/java/com/so_tro_online/quan_ly_hoa_don/dto/ChiTietHoaDonResponse.java
@@ -11,12 +11,10 @@ import java.math.BigDecimal;
 
 public class ChiTietHoaDonResponse {
     private Integer id;
-
     private Integer maHoaDon;
     private String tenDichVu;
-
     private BigDecimal donGia;
-    private BigDecimal soLuong;
+    private Integer soLuong;
     private BigDecimal thanhTien;
     private BigDecimal heSo;
     private BigDecimal tienThucTe;
@@ -54,11 +52,11 @@ public class ChiTietHoaDonResponse {
         this.donGia = donGia;
     }
 
-    public BigDecimal getSoLuong() {
+    public Integer getSoLuong() {
         return soLuong;
     }
 
-    public void setSoLuong(BigDecimal soLuong) {
+    public void setSoLuong(Integer soLuong) {
         this.soLuong = soLuong;
     }
